Added cron, added documentation
