+++ etc/crontab.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Cron:etc/crontab.xsd">
+<group id="slack_cron">
+    <job name="slack" instance="Pimgento\Api\Cron\SlackNotificationCron" method="execute">
+        <schedule>0 8 * * *</schedule>
+    </job>
+</group>
+</config>
